Add online download UI and catalog loading
Introduce online download feature for the title scene: add new source files for catalog loading (online_download_catalog.cpp), rendering (online_download_render.cpp), and input/update handling (online_download_update.cpp), plus an internal header (online_download_internal.h). Implement local song discovery, jacket caching, async catalog reload/polling, song/chart state building, and comprehensive UI rendering and interaction logic. Wire new files into CMakeLists.txt and refactor online_download_view.cpp to use the internal header and remove duplicated includes. Note: remote catalog loading is stubbed out (load_remote_song_entries returns empty) and can be implemented later.
